check all builidentities instead of exiting after failure on first one
Previous behavior was to exit out of finding buildidentities appropriate for the tss request when the first error is encoutered. But the actually appropirate buildidentity may not contain any errors and might be a few indexes after the one that caused the error.
Now, if there's an error on builidentity, it skips the buildidentity. A proper buildidentity to return is only set up on it's exactly matched.